# Enable key based caching for Danger