--- a/.build/workspace-state.json
+++ b/.build/workspace-state.json
@@ -1,0 +1,143 @@
+{
+  "object" : {
+    "artifacts" : [
+
+    ],
+    "dependencies" : [
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "files",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/JohnSundell/Files",
+          "name" : "Files"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "d273b5b7025d386feef79ef6bad7de762e106eaf",
+            "version" : "4.2.0"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "Files"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "logger",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/shibapm/Logger",
+          "name" : "Logger"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "53c3ecca5abe8cf46697e33901ee774236d94cce",
+            "version" : "0.2.3"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "Logger"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "octokit.swift",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/nerdishbynature/octokit.swift",
+          "name" : "OctoKit"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "9521cdff919053868ab13cd08a228f7bc1bde2a9",
+            "version" : "0.11.0"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "octokit.swift"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "requestkit",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/nerdishbynature/RequestKit.git",
+          "name" : "RequestKit"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "fd5e9e99aada7432170366c9e95967011ce13bad",
+            "version" : "2.4.0"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "RequestKit"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "swift",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/danger/swift",
+          "name" : "danger-swift"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "c1bba33f705ca0fd4a0022997c8b696210083755",
+            "version" : "3.12.3"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "swift"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "version",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/mxcl/Version",
+          "name" : "Version"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "200046c93f6d5d78a6d72bfd9c0b27a95e9c0a2b",
+            "version" : "1.2.0"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "Version"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "wetransferprlinter",
+          "kind" : "fileSystem",
+          "location" : "/Users/avanderlee/Developer/GIT-Projects/WeTransfer/Mule/Submodules/WeTransfer-iOS-SDK/Submodules/WeTransfer-iOS-CI/WeTransferPRLinter",
+          "name" : "WeTransferPRLinter"
+        },
+        "state" : {
+          "name" : "fileSystem",
+          "path" : "/Users/avanderlee/Developer/GIT-Projects/WeTransfer/Mule/Submodules/WeTransfer-iOS-SDK/Submodules/WeTransfer-iOS-CI/WeTransferPRLinter"
+        },
+        "subpath" : "wetransferprlinter"
+      },
+      {
+        "basedOn" : null,
+        "packageRef" : {
+          "identity" : "xcresultkit",
+          "kind" : "remoteSourceControl",
+          "location" : "https://github.com/davidahouse/XCResultKit.git",
+          "name" : "XCResultKit"
+        },
+        "state" : {
+          "checkoutState" : {
+            "revision" : "2fe24c2675432c4b8c8d7ab0364e2760190b6fea",
+            "version" : "0.9.3"
+          },
+          "name" : "sourceControlCheckout"
+        },
+        "subpath" : "XCResultKit"
+      }
+    ]
+  },
+  "version" : 5
+}