--- a/Bitrise/testing_bitrise.yml
+++ b/Bitrise/testing_bitrise.yml
@@ -119,6 +119,9 @@ workflows:
         - is_compress: 'true'
         - debug_mode: 'true'
         - deploy_path: build/derived_data/Build/Products/Debug-iphonesimulator/${XCODE_TARGET}.app/  
+    - restore-cache@1:          
+        inputs:          
+        - key: danger-build-cache
     - script:
         title: Run Danger
         deps:
@@ -139,11 +142,16 @@ workflows:
 
             if [ "$BITRISEIO_GIT_REPOSITORY_SLUG" == "WeTransfer-iOS-CI" ]; then
                 chmod +x danger-swift
-                ./danger-swift ci --cache-path .build --build-path .build
+                ./danger-swift ci --cache-path .danger-build --build-path .danger-build
             else
                 cd Submodules/WeTransfer-iOS-CI
-                ./danger-swift ci --cwd ../../ --cache-path ../../.build --build-path ../../.build
+                ./danger-swift ci --cwd ../../ --cache-path .danger-build --build-path .danger-build
             fi
+    - save-cache@1:
+        inputs:
+        - key: danger-build-cache
+        - paths: .danger-build
+        - is_key_unique: 'false'
     - deploy-to-bitrise-io@2:
         is_skippable: true
         inputs: