--- a/Bitrise/testing_bitrise.yml
+++ b/Bitrise/testing_bitrise.yml
@@ -66,8 +66,8 @@ workflows:
         - content: |-
             #!/usr/bin/env bash
             # Use the environment FASTLANE_LANE if available. Otherwise, fallback to "test"
-            lane=${FASTLANE_LANE:=test}
-            fastlane $lane
+            #lane=${FASTLANE_LANE:=test}
+            #fastlane $lane
     - deploy-to-bitrise-io@2:
         title: Deploy Simulator Build to Bitrise
         is_skippable: true
@@ -79,7 +79,7 @@ workflows:
         - deploy_path: build/derived_data/Build/Products/Debug-iphonesimulator/${XCODE_TARGET}.app/  
     - restore-cache@1:          
         inputs:          
-        - key: danger-build-cache-{{ checksum ".build/workspace-state.json" }}
+        - key: danger-build-cache-{{ checksum ".build/workspace-state.json" "Submodules/WeTransfer-iOS-CI/.build/workspace-state.json" }}
     - script:
         title: Run Danger
         deps:
@@ -101,14 +101,16 @@ workflows:
             if [ "$BITRISEIO_GIT_REPOSITORY_SLUG" == "WeTransfer-iOS-CI" ]; then
                 chmod +x danger-swift
                 ./danger-swift ci
+                envman add --key DANGER_BUILD_DIRECTORY --value ".build"
             else
                 cd Submodules/WeTransfer-iOS-CI
                 ./danger-swift ci --cwd ../../
+                envman add --key DANGER_BUILD_DIRECTORY --value "Submodules/WeTransfer-iOS-CI/.build"
             fi
     - save-cache@1:
         inputs:
-        - key: danger-build-cache-{{ checksum ".build/workspace-state.json" }}
-        - paths: .build
+        - key: danger-build-cache-{{ checksum ".build/workspace-state.json" "Submodules/WeTransfer-iOS-CI/.build/workspace-state.json" }}
+        - paths: "$DANGER_BUILD_DIRECTORY"
         - is_key_unique: 'true'
     - deploy-to-bitrise-io@2:
         is_skippable: true