Caching the `.build` directory for Danger using key-based caching will speed up CI runs. Caching is based on the dependencies used for Danger and will survive between different branches.

| Step | Before | After |
| --- | --- | --- |
| Run Danger | 66s | 39s |