@wetransferplatform 2022-11-22
<!--
  0 failure: 
  0 warning: 
  2 messages
  1 markdown notices
  DangerID: danger-id-Danger;
-->



<table>
  <thead>
    <tr>
      <th width="50"></th>
      <th width="100%" data-danger-table="true">Messages</th>
    </tr>
  </thead>
  <tbody><tr>
      <td>:book:</td>
      <td>

  View more details on <a href="https://app.bitrise.io/build/fd7fbb85-df70-4935-9426-cade73287311" target="_blank">Bitrise</a>
  </td>
    </tr>
  
<tr>
      <td>:book:</td>
      <td>WeTransferPRLinterTests: Executed 19 tests (0 failed, 0 retried, 0 skipped) in 10.053 seconds</td>
    </tr>
  </tbody>
</table>

## Code Coverage Report
| Name | Coverage ||
| --- | --- | --- |
WeTransferPRLinter | 94.25% | ✅

<p align="right">
  Generated by :no_entry_sign: <a href="https://danger.systems/swift">Danger Swift</a> against bef7be8e4c033e264262cdd01bcba4128294223c
</p>